+++ lib/kapnismology/engine.rb
@@ -3,10 +3,10 @@ module Kapnismology
   begin
     # Rails engine to automatically load our code and smoke tests libraries
     class Engine < ::Rails::Engine
-      initializer 'kapnismology.add_autoload_paths', before: :set_autoload_paths do |app|
+      initializer 'kapnismology.add_autoload_paths', before: :set_autoload_paths do |_app|
         smoketest_dir = Rails.root.join('lib', 'smoke_test')
         if smoketest_dir.exist?
-          Dir[File.join(smoketest_dir, '*.rb')].each {|file| require file }
+          Dir[File.join(smoketest_dir, '*.rb')].each { |file| require file }
         end
       end