+++ spec/internal/config/routes.rb
@@ -1,3 +1,5 @@
+require 'kapnismology/engine'
+
 Rails.application.routes.draw do
   mount Kapnismology::Engine, at: '/smoke_test'
 end