+++ spec/lib/kapnismology/result_spec.rb
@@ -1,20 +1,42 @@
 require 'spec_helper'
 
-module Kapnismology
-  RSpec.describe Result do
-    let(:data) { [:berserk] }
-    let(:passed) { true }
-    let(:message) { '黄金時代' }
-    let(:result) { Result.new(passed, data, message) }
+RSpec.describe Kapnismology::Result do
+  let(:data) { [:berserk] }
+  let(:passed) { true }
+  let(:message) { '黄金時代' }
+  let(:name) { 'Kentaro Miura' }
+  let(:result) { described_class.new(passed, data, message) }
+  let(:extra_char) { '' }
 
-    it 'provide access to data, passed and message' do
-      expect(result.data).to eq(data)
-      expect(result.passed?).to eq(passed)
-      expect(result.message).to eq(message)
+  shared_examples_for 'serializes its data' do
+    it 'creates a string with its data' do
+      expected = "\e[32m\e[1mpassed\e[0m: #{name}\n#{extra_messages.join("\n")}#{extra_char}\e[1m#{message}\e[0m\n   #{data}\n"
+      expect(result.to_s(name)).to eq(expected)
     end
 
     it 'creates a hash with its data' do
-      expect(result.to_hash).to eq(passed: passed, data: data, message: message)
+      expected = { passed: passed, data: data, message: message, extra_messages: extra_messages }
+      expect(result.to_hash).to eq(expected)
     end
   end
+
+  it 'inherits from BaseResult' do
+    expect(result.is_a?(Kapnismology::BaseResult)).to eq(true)
+  end
+
+  context 'no extra messages added' do
+    let(:extra_messages) { [] }
+    it_behaves_like 'serializes its data'
+  end
+
+  context 'added some extra messages' do
+    let(:extra_messages) { ['42', '41'] }
+    let(:extra_char) { "\n" }
+
+    before do
+      result.add_extra_messages(['42', '41'])
+    end
+    it_behaves_like 'serializes its data'
+  end
+
 end